Add Router + Add possibility to add TX has from URL
